requirements.txt
openai>=1.30.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langdetect>=1.0.9
razdel>=0.5.0
